app/requirements.txt
aiofiles==0.5.0
appdirs==1.4.4
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
cryptography==3.1.1
dnspython==2.0.0
ecdsa==0.14.1
email-validator==1.1.1
fastapi==0.61.1
fastapi-jwt-auth==0.1.0
fastapi-login==1.5.1
fastapi-users==3.0.5
h11==0.9.0
httptools==0.1.1
idna==2.10
inflect==4.1.0
makefun==1.9.3
mccabe==0.6.1
mypy-extensions==0.4.3
passlib==1.7.2
pathspec==0.8.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.6.1
pyflakes==2.2.0
PyJWT==1.7.1
python-jose==3.2.0
python-multipart==0.0.5
regex==2020.7.14
rsa==4.6
six==1.15.0
sqlacodegen==2.3.0
SQLAlchemy==1.3.19
SQLAlchemy-serializer==1.3.4.2
starlette==0.13.6
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
